Refactor to conventions
